Add user_language to execution environment

+++ library/persistency/pyxis-doc/parser/el_pyxis_to_xml_converter.e
@@ -6,8 +6,8 @@ note
 	contact: "finnian at eiffel hyphen loop dot com"
 
 	license: "MIT license (See: en.wikipedia.org/wiki/MIT_License)"
-	date: "2022-02-08 10:28:52 GMT (Tuesday 8th February 2022)"
-	revision: "11"
+	date: "2022-03-01 16:34:43 GMT (Tuesday 1st March 2022)"
+	revision: "12"
 
 class
 	EL_PYXIS_TO_XML_CONVERTER
@@ -17,7 +17,7 @@ inherit
 
 	EL_MODULE_LIO
 
-	EL_MODULE_FILE_SYSTEM
+	EL_MODULE_FILE; EL_MODULE_FILE_SYSTEM
 
 	EL_FILE_OPEN_ROUTINES
 
@@ -32,8 +32,6 @@ feature {EL_COMMAND_CLIENT} -- Initialization
 			create source_encoding.make_from_file (source_path)
 			if output_path.is_empty then
 				output_path := new_output_path
-			else
-				File_system.make_directory (a_output_path.parent)
 			end
 			xml_generator := new_xml_generator
 		end
@@ -58,19 +56,27 @@ feature -- Basic operations
 		local
 			in_file, out_file: PLAIN_TEXT_FILE
 		do
-			if is_lio_enabled then
-				lio.put_path_field ("Converting " + source_encoding.name, source_path)
+
+			if File.is_newer_than (source_path, output_path) then
+				if is_lio_enabled then
+					lio.put_path_field ("Converting " + source_encoding.name, source_path)
+					lio.put_new_line
+				end
+				if output_path.exists then
+					File_system.remove_file (output_path)
+				else
+					File_system.make_directory (output_path.parent)
+				end
+				create out_file.make_open_write (output_path)
+				create in_file.make_open_read (source_path)
+
+				xml_generator.convert_stream (in_file, out_file)
+				out_file.close; in_file.close
+				
+			elseif is_lio_enabled then
+				lio.put_path_field ("No change in %S", source_path)
 				lio.put_new_line
 			end
-
-			if output_path.exists then
-				File_system.remove_file (output_path)
-			end
-			create out_file.make_open_write (output_path)
-			create in_file.make_open_read (source_path)
-
-			xml_generator.convert_stream (in_file, out_file)
-			out_file.close; in_file.close
 		end
 
 feature {NONE} -- Implementation
